% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\encoding{UTF-8}
\name{available_templates}
\alias{available_templates}
\title{available_templates}
\usage{
available_templates()
}
\value{
a character list with the names of the built-in tepmlates available 
in the package
}
\description{
available_templates
}
\examples{
available_templates()

}
