% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\encoding{UTF-8}
\name{check_node}
\alias{check_node}
\title{check_node}
\usage{
check_node(x)
}
\arguments{
\item{x}{\\code{list}, the node to check}
}
\value{
if the node is ok, the node; else, an error is raised.
}
\description{
Verify node integrity while navigating the structure tree
}
\examples{
rprojtree:::check_node(rjson::fromJSON('{"type":"file", "name": "global.R"}'))

}
\keyword{internal}
