% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprojtree-package.R
\docType{package}
\name{rprojtree-package}
\alias{rprojtree-package}
\alias{_PACKAGE}
\alias{rprojtree}
\title{rprojtree}
\description{
This package allows you to easily create file and directory structures
for data science projects
}
\details{
For more details see the tutorial vignette:

\itemize{
  \item \code{vignette("rprojtree-primer", package = "rprojtree")}
}
}
