% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make_prj_tree}
\alias{make_prj_tree}
\title{make_prj_tree}
\usage{
make_prj_tree(json_str, file, path = ".", verbose = FALSE)
}
\arguments{
\item{json_str}{asdfasdfasdf}

\item{file}{asdfasdf}

\item{path}{asdfasdf}

\item{verbose}{afsdgasdf}
}
\value{

}
\description{
make_prj_tree
}
\details{

}
\examples{
\dontrun{
make_prj_tree(json_str = "basic_template", path = "c:/Users/migue/Documents/tmp/rprj_tst/")

make_prj_tree(file = "C:/Users/migue/OneDrive/Documentos/rprj_tree_tst.json", path = "c:/Users/migue/Documents/tmp/rprj_tst/")
}

}
