% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make_prj_tree}
\alias{make_prj_tree}
\title{make_prj_tree}
\usage{
make_prj_tree(json_str, file, path = ".", verbose = FALSE)
}
\arguments{
\item{json_str}{}

\item{file}{}

\item{path}{}

\item{verbose}{}
}
\value{

}
\description{
make_prj_tree
}
