% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{print_template}
\alias{print_template}
\title{print_template}
\usage{
print_template(template_name)
}
\arguments{
\item{template_name}{}
}
\value{

}
\description{
print_template
}
\examples{
print_template("basic_template")

}
