% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\encoding{UTF-8}
\name{print_template}
\alias{print_template}
\title{print_template}
\usage{
print_template(template_name)
}
\arguments{
\item{template_name}{name of the builtin template or the .json file to use}
}
\description{
print the files structure described by a builtin template or by a .json file
}
\examples{
print_template("basic_template")

}
